requirements.txt
pypdf>=4.2.0
google-generativeai>=0.7.2

python-docx>=1.1.0
fastapi>=0.112.0
uvicorn>=0.30.0
